Set a user as a link parent when creating route

+++ src/main/java/com/google/sps/data/Marker.java
@@ -19,13 +19,13 @@ public class Marker {
   private double lat;
   private double lng;
 
-  private int stayHour;
-  private int stayMinute;
+  private long stayHour;
+  private long stayMinute;
 
   private String markerName;
 
   public Marker(
-      double newLat, double newLng, int newStayHour, int newStayMinute, String newMarkerName) {
+      double newLat, double newLng, long newStayHour, long newStayMinute, String newMarkerName) {
     lat = newLat;
     lng = newLng;
 
@@ -53,11 +53,11 @@ public class Marker {
     return lng;
   }
 
-  public int getStayHour() {
+  public long getStayHour() {
     return stayHour;
   }
 
-  public int getStayMinute() {
+  public long getStayMinute() {
     return stayMinute;
   }
 
@@ -73,11 +73,11 @@ public class Marker {
     lng = newLng;
   }
 
-  public void setStayHour(int newHour) {
+  public void setStayHour(long newHour) {
     stayHour = newHour;
   }
 
-  public void setStayMinute(int newMinute) {
+  public void setStayMinute(long newMinute) {
     stayMinute = newMinute;
   }
